+++ models/user_meta.js
@@ -3,13 +3,13 @@
 module.exports = function(sequelize, DataTypes) {
   var user_meta = sequelize.define('user_meta', {
     ID: {
-      type: DataTypes.INTEGER(11).UNSIGNED,
+      type: DataTypes.INTEGER.UNSIGNED,
       allowNull: false,
       primaryKey: true,
       autoIncrement: true
     },
     user_id: {
-      type: DataTypes.INTEGER(11).UNSIGNED,
+      type: DataTypes.INTEGER.UNSIGNED,
       allowNull: false,
       references: {
         model: 'user',
